Made video.paused and video.start() non-blocking
video.paused is now a regular attribute that gets updated from an event
handler. Also, video.start() no longer busy waits until the video is
ready to seek to the requested start position; it sets a request that is
also handled in an event handler.

These changes mean Fabella is no longer blocked for hundreds of ms on
certain calls after video start, solving the severe animation stutters
that occured on video start.